Replace float -> double in prices
The problem lies within qml, somehow float properties converted between
qml and C++ cause strange rounding effects. This hower is not the case
for double properties.

+++ src/domain/nina/Position.cpp
@@ -13,8 +13,8 @@ Position::Position()
 Position::Position(
     const std::string& description,
     const std::string& unit,
-    float              pricePerUnit,
-    float              numUnits
+    double             pricePerUnit,
+    double             numUnits
     )
     : m_description( description )
     , m_unit( unit )
@@ -29,7 +29,7 @@ Position::Position(
 
 Position::Position(
     const std::string& description,
-    float              price
+    double             price
     )
     : m_description( description )
     , m_pricePerUnit( 0 )
